ButtonMyList component created
