+++ project/src/const.ts
@@ -93,6 +93,11 @@ export enum Rating {
   Undefined = '',
 }
 
+export enum MyListIcon {
+  Favorite = '#in-list',
+  NotFavorite = '#add',
+}
+
 export const RATING_DIGITS = 1;
 
 export const MONTHS = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'];
